requirements.txt
# Movie Notifier Dependencies
requests>=2.31.0
PyYAML>=6.0
python-dotenv>=1.0.0
schedule>=1.2.0
pytz>=2023.3
tzlocal>=5.0.1
colorama>=0.4.6  # For colored console output